# docs/api.yaml
openapi: 3.0.3
info:
  title: Love-in-hg 相亲角 API
  version: 1.0.0
  description: |
    相亲角小程序后端接口文档。
    - 生产环境: `https://aibao.love/api`
    - 本地开发: `http://127.0.0.1:5000/api`
    - 所有响应统一格式: `{ code: number, msg: string, data: any }`
    - 需要登录的接口请在请求头带上 `X-User-Id`

servers:
  - url: https://aibao.love/api
    description: 生产环境
  - url: http://127.0.0.1:5000/api
    description: 本地开发

tags:
  - name: 认证
    description: 登录相关
  - name: 启事
    description: 启事发布、查看、分配
  - name: 用户
    description: 用户信息
  - name: 会员
    description: 会员购买
  - name: 管理
    description: 管理接口（数据导入）
  - name: 健康检查
    description: 服务状态

paths:
  /../ping:
    get:
      tags: [健康检查]
      summary: 健康检查
      description: 检查后端服务是否正常运行
      responses:
        '200':
          description: 服务正常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'

  /../:
    get:
      tags: [健康检查]
      summary: 根路径
      description: 后端服务状态信息
      responses:
        '200':
          description: 服务信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  msg: { type: string, example: Love-in-hg backend is running }

  /auth/login:
    post:
      tags: [认证]
      summary: 手机号登录
      description: |
        弱校验登录：只要启事表中存在该手机号即可登录。
        若用户不存在会自动创建，数据从对应的启事中提取。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  pattern: '^1[3-9]\d{9}$'
                  description: 11位手机号
                  example: '13800138001'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Ok'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/User'
                          isNewUser:
                            type: boolean
                            example: false
        '400':
          description: 手机号格式错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 400
                msg: 请输入正确的手机号码

  /notice/publish:
    post:
      tags: [启事]
      summary: 发布启事
      description: |
        一个手机号可发布多条启事。
        若用户不存在会自动创建。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoticeInput'
      responses:
        '200':
          description: 发布成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Ok'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          noticeId:
                            type: string
                            example: '101'
                          userId:
                            type: string
                            example: '1'
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /notice/update/{noticeId}:
    post:
      tags: [启事]
      summary: 修改启事
      description: 修改已有启事信息
      parameters:
        - in: path
          name: noticeId
          required: true
          schema:
            type: integer
          example: 101
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoticeInput'
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Ok'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          noticeId: { type: string, example: '101' }
        '404':
          description: 启事不存在

  /notice/delete/{noticeId}:
    post:
      tags: [启事]
      summary: 删除启事
      description: 仅启事所有者可删除，会同时删除关联的浏览记录
      parameters:
        - in: path
          name: noticeId
          required: true
          schema:
            type: integer
          example: 101
      responses:
        '200':
          description: 删除成功
        '401':
          description: 未登录
        '403':
          description: 无权删除他人启事

  /notice/list:
    get:
      tags: [启事]
      summary: 获取分配的启事列表
      description: |
        按月/按天分配新启事（排除历史已分配过的），
        已分配的启事在周期内可反复查看，不消耗配额。
      parameters:
        - in: query
          name: userId
          schema: { type: string }
          description: 用户ID（也可通过 X-User-Id 头传递）
        - in: header
          name: X-User-Id
          schema: { type: string }
          description: 用户ID
        - in: query
          name: gender
          schema: { type: string, enum: ['男', '女'] }
          description: 性别筛选
        - in: query
          name: minHeight
          schema: { type: integer }
          description: 最低身高 cm
        - in: query
          name: maxHeight
          schema: { type: integer }
          description: 最高身高 cm
        - in: query
          name: minAge
          schema: { type: integer }
          description: 最小年龄
        - in: query
          name: maxAge
          schema: { type: integer }
          description: 最大年龄
        - in: query
          name: income
          schema: { type: string }
          description: 收入筛选
        - in: query
          name: isPublicSector
          schema: { type: string, enum: ['true', 'false'] }
          description: 是否体制内
        - in: query
          name: occupation
          schema: { type: string }
          description: 职业关键词
      responses:
        '200':
          description: 启事列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Ok'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          notices:
                            type: array
                            items:
                              $ref: '#/components/schemas/NoticeMasked'
                          remaining:
                            type: integer
                            example: 10
                          membershipType:
                            type: string
                            enum: [free, member, vip]
                          periodType:
                            type: string
                            enum: [monthly, daily]
                          limitReached:
                            type: boolean
        '401':
          description: 未登录

  /notice/history:
    get:
      tags: [启事]
      summary: 历史查看
      description: 系统分配给该用户的所有启事（跨周期汇总）
      parameters:
        - in: query
          name: userId
          schema: { type: string }
        - in: header
          name: X-User-Id
          schema: { type: string }
      responses:
        '200':
          description: 历史列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Ok'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          notices:
                            type: array
                            items:
                              $ref: '#/components/schemas/NoticeMasked'

  /notice/viewPhone:
    post:
      tags: [启事]
      summary: 查看完整手机号
      description: 按需获取启事的完整手机号（列表返回的是脱敏号）
      parameters:
        - in: query
          name: userId
          schema: { type: string }
        - in: header
          name: X-User-Id
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [noticeId]
              properties:
                noticeId:
                  type: integer
                  example: 101
      responses:
        '200':
          description: 完整手机号
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Ok'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          noticeId: { type: string, example: '101' }
                          name: { type: string, example: 张三 }
                          phone: { type: string, example: '13800138001' }

  /user/profile:
    get:
      tags: [用户]
      summary: 获取用户信息
      description: 用户基本信息 + 配额
      parameters:
        - in: query
          name: userId
          schema: { type: string }
        - in: query
          name: phone
          schema: { type: string }
        - in: header
          name: X-User-Id
          schema: { type: string }
      responses:
        '200':
          description: 用户信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Ok'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            allOf:
                              - $ref: '#/components/schemas/User'
                              - type: object
                                properties:
                                  noticeCount:
                                    type: integer
                                    description: 已发布启事数量
                                    example: 2
                          quota:
                            $ref: '#/components/schemas/Quota'

  /user/notices:
    get:
      tags: [用户]
      summary: 我的启事
      description: 当前用户发布的所有启事
      parameters:
        - in: query
          name: userId
          schema: { type: string }
        - in: header
          name: X-User-Id
          schema: { type: string }
      responses:
        '200':
          description: 启事列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Ok'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          notices:
                            type: array
                            items:
                              $ref: '#/components/schemas/Notice'

  /user/deleteAccount:
    post:
      tags: [用户]
      summary: 注销账号
      description: 删除用户及其全部关联数据（启事、浏览记录、会员订单）
      parameters:
        - in: query
          name: userId
          schema: { type: string }
        - in: header
          name: X-User-Id
          schema: { type: string }
      responses:
        '200':
          description: 注销成功
        '401':
          description: 未登录

  /membership/purchase:
    post:
      tags: [会员]
      summary: 创建会员订单
      description: |
        创建会员购买订单（模拟，实际需对接微信支付）
        - member（会员）: 99元
        - vip（大会员）: 999元
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                  enum: [member, vip]
                  example: member
                userId:
                  type: string
                  description: 可选，也可通过 X-User-Id 头传递
      responses:
        '200':
          description: 订单创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Ok'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          orderId: { type: string, example: '1' }
                          amount: { type: integer, example: 99 }
                          type: { type: string, example: member }
                          typeLabel: { type: string, example: 会员 }
        '400':
          description: 无效的会员类型

  /admin/extractAndImport:
    post:
      tags: [管理]
      summary: 文本提取导入
      description: |
        从长文本中提取键值对并入库。
        支持格式: `{姓名：张三}` / `姓名：张三` / `姓名:张三`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  example: '{姓名：李四，性别：女，出生年月：1995-06-15，身高：165，体重：50，职业：教师，收入：15-20万，是否体制内：是，手机号：13900139001}'
      responses:
        '200':
          description: 导入成功
        '409':
          description: 手机号已存在

  /admin/initData:
    post:
      tags: [管理]
      summary: 批量导入种子数据
      description: 读取 sql/seed_notices.json 批量导入。已存在的手机号自动跳过。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  description: 管理密码（默认 admin123）
                  example: admin123
      responses:
        '200':
          description: 导入完成
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Ok'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          successCount: { type: integer, example: 100 }
                          skipCount: { type: integer, example: 0 }
                          errorCount: { type: integer, example: 0 }
        '403':
          description: 密码错误

components:
  schemas:
    Ok:
      type: object
      properties:
        code: { type: integer, example: 0 }
        msg: { type: string, example: 成功 }
      required: [code, msg]

    Error:
      type: object
      properties:
        code: { type: integer, example: 400 }
        msg: { type: string, example: 错误信息 }
      required: [code, msg]

    User:
      type: object
      properties:
        id: { type: integer, example: 1 }
        phone: { type: string, example: '13800138001' }
        name: { type: string, example: 张三 }
        gender: { type: string, example: 男 }
        birthday: { type: string, example: '1990-03-15' }
        occupation: { type: string, example: 互联网 }
        income: { type: string, example: '20-30万' }
        isPublicSector: { type: boolean, example: false }
        height: { type: integer, example: 178 }
        weight: { type: integer, example: 65 }
        membershipType:
          type: string
          enum: [free, member, vip]
          example: free
        membershipExpire:
          type: string
          nullable: true
          example: null
        createdAt: { type: string, example: '2026-09-01T08:00:00' }

    NoticeInput:
      type: object
      required: [phone, name, gender, birthday]
      properties:
        phone:
          type: string
          pattern: '^1[3-9]\d{9}$'
          example: '13800138001'
        name:
          type: string
          example: 张三
        nickname:
          type: string
          example: 小张
        gender:
          type: string
          enum: [男, 女]
          example: 男
        birthday:
          type: string
          format: date
          example: '1990-03-15'
        occupation:
          type: string
          example: 互联网
        income:
          type: string
          example: '20-30万'
        isPublicSector:
          type: boolean
          example: false
        height:
          type: integer
          example: 178
        weight:
          type: integer
          example: 65
        remark:
          type: string
          example: 性格开朗，喜欢旅行

    Notice:
      type: object
      properties:
        id: { type: integer, example: 1 }
        _id: { type: string, example: '1' }
        userId: { type: string, example: '1' }
        phone: { type: string, example: '13800138001' }
        name: { type: string, example: 张三 }
        nickname: { type: string, example: 小张 }
        gender: { type: string, example: 男 }
        birthday: { type: string, example: '1990-03-15' }
        occupation: { type: string, example: 互联网 }
        income: { type: string, example: '20-30万' }
        isPublicSector: { type: boolean, example: false }
        height: { type: integer, example: 178 }
        weight: { type: integer, example: 65 }
        source:
          type: string
          enum: [user, seed, import]
          example: seed
        remark: { type: string, example: '' }
        createdAt: { type: string, example: '2026-09-01T08:00:00' }

    NoticeMasked:
      allOf:
        - $ref: '#/components/schemas/Notice'
        - type: object
          description: 手机号脱敏版本（138****0001）
          properties:
            phone:
              type: string
              example: '138****0001'

    Quota:
      type: object
      properties:
        membershipType:
          type: string
          enum: [free, member, vip]
          example: free
        periodType:
          type: string
          enum: [monthly, daily]
          description: free/member=monthly, vip=daily
          example: monthly
        monthlyLimit:
          type: integer
          example: 10
          description: 配额上限（free=10, member=30, vip=30）
        monthlyAssigned:
          type: integer
          example: 10
          description: 本周期已分配数量
        remaining:
          type: integer
          example: 10
          description: 等于 monthlyAssigned（分配后可反复查看）
        limitReached:
          type: boolean
          example: false